Moved site content widget components to 'widgets' folder
