+++ static/js/components/SiteContentListing.test.tsx
@@ -33,7 +33,7 @@ jest.mock("react-router-dom", () => ({
 }))
 
 // ckeditor is not working properly in tests, but we don't need to test it here so just mock it away
-jest.mock("./MarkdownEditor", () => "div")
+jest.mock("./widgets/MarkdownEditor", () => "div")
 
 describe("SiteContentListing", () => {
   let helper: IntegrationTestHelper,
