+++ static/js/pages/MarkdownEditorTestPage.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react"
 import { TabContent, TabPane, Nav, NavItem, NavLink } from "reactstrap"
 
-import MarkdownEditor from "../components/MarkdownEditor"
+import MarkdownEditor from "../components/widgets/MarkdownEditor"
 
 import { TEST_MARKDOWN } from "../test_constants"
